verify pending asset module added and integrated in setting tab and record not found handled in settings
