+++ src/app/dashboard/dashboard.component.html
@@ -82,10 +82,6 @@
             <p class="graph-font">{{MonthlyAssetTitle}}-Year ({{assetselectedyear}})</p>
           </div>
           <div fxLayout="row" fxLayoutAlign="end end">
-            <!-- <button class="fab-btn-dashboard" matTooltip="print graph" mat-mini-fab color="primary"
-              (click)="printGraph('monthlyasset', MonthlyAssetTitle +' - Year'+' ('+assetselectedyear+')')">
-              <mat-icon color="primary">print</mat-icon>
-            </button> -->
             <button class="fab-btn-dashboard" mat-icon-button [matMenuTriggerFor]="menu">
               <mat-icon>more_vert</mat-icon>
             </button>
@@ -116,10 +112,6 @@
             <p class="graph-font">{{MonthlyComplaintsTitle}}-Year ({{complaintselectedyear}})</p>
           </div>
           <div fxLayout="row" fxLayoutAlign="end end">
-            <!-- <button class="fab-btn-dashboard" matTooltip="print graph" mat-mini-fab color="primary"
-              (click)="printGraph('monthlycomplaints', MonthlyComplaintsTitle +' - Year'+' ('+complaintselectedyear+')')">
-              <mat-icon color="primary">print</mat-icon>
-            </button> -->
             <button class="fab-btn-dashboard" mat-icon-button [matMenuTriggerFor]="menu1">
               <mat-icon>more_vert</mat-icon>
             </button>
@@ -148,10 +140,6 @@
     fxLayoutAlign="space-around center" fxLayoutGap="30px">
     <div fxFlex.xs="100" fxFlex="50">
       <mat-card>
-        <!-- <button class="fab-btn-dashboard" matTooltip="print graph" mat-mini-fab color="primary"
-          (click)="printGraph('CatAsset', CategoryAssetTitle)">
-          <mat-icon color="primary">print</mat-icon>
-        </button> -->
         <button class="fab-btn-dashboard" mat-icon-button [matMenuTriggerFor]="menu2">
           <mat-icon>more_vert</mat-icon>
         </button>
@@ -169,10 +157,6 @@
     </div>
     <div fxFlex.xs="100" fxFlex="50">
       <mat-card>
-        <!-- <button class="fab-btn-dashboard" matTooltip="print graph" mat-mini-fab color="primary"
-          (click)="printGraph('Pendingmaintainance', CategoryMaintainanceTitle)">
-          <mat-icon color="primary">print</mat-icon>
-        </button> -->
         <button class="fab-btn-dashboard" mat-icon-button [matMenuTriggerFor]="menu3">
           <mat-icon>more_vert</mat-icon>
         </button>
@@ -190,13 +174,9 @@
     </div>
   </div>
   <div class="graph-m-t" fxLayout="row" fxLayout.xs="column" fxLayoutAlign.xs="space-around"
-    fxLayoutAlign="space-between start" fxLayoutGap="30px">
+    fxLayoutAlign="center center" fxLayoutGap="30px">
     <div fxFlex.xs="100" fxFlex="50">
       <mat-card>
-        <!-- <button class="fab-btn-dashboard" matTooltip="print graph" mat-mini-fab color="primary"
-          (click)="printGraph('installationlocation', InstallLocAssetTitle)">
-          <mat-icon color="primary">print</mat-icon>
-        </button> -->
         <button class="fab-btn-dashboard" mat-icon-button [matMenuTriggerFor]="menu4">
           <mat-icon>more_vert</mat-icon>
         </button>
@@ -210,55 +190,6 @@
           <p class="graph-font">{{InstallLocAssetTitle}}</p>
           <canvas id="installationlocation" width="40" height="20"></canvas>
         </div>
-      </mat-card>
-    </div>
-    <div fxFlex.xs="100" fxFlex="50">
-      <mat-card>
-        <div fxLayout="column" fxLayoutAlign="center center" *ngIf="data4">
-          <p class="graph-font">VERIFICATION PENDING - ASSETS</p>
-        </div>
-        <div *ngIf="isNoRecordFound; else recordFound" class="no-record-found-container">
-          <p>No Record Found</p>
-        </div>
-        <ng-template #recordFound>
-          <mat-table [dataSource]="dataSource">
-            <ng-container matColumnDef="assetTitle">
-              <mat-header-cell fxLayoutAlign="start center" *matHeaderCellDef>Asset Title
-              </mat-header-cell>
-              <mat-cell fxLayoutAlign="start center" *matCellDef="let visit">
-                {{visit.assetTitle}}
-              </mat-cell>
-            </ng-container>
-            <ng-container matColumnDef="addedBy">
-              <mat-header-cell fxHide.xs fxLayoutAlign="center center" *matHeaderCellDef>Added By
-              </mat-header-cell>
-              <mat-cell fxHide.xs fxLayoutAlign="center center" *matCellDef="let visit">
-                {{visit.addedBy}}
-              </mat-cell>
-            </ng-container>
-            <ng-container matColumnDef="action">
-              <mat-header-cell fxLayoutAlign="end center" *matHeaderCellDef></mat-header-cell>
-              <mat-cell (click)="$event.stopPropagation()" *matCellDef="let visit; let i = index;" flexDirection="row"
-                fxLayoutAlign="end center" fxLayoutGap="20px">
-                <mat-icon matTooltip="mark as verified" class="mat-ic"
-                  (click)="$event.stopPropagation(); verifiedAsset(visit.assetId);">
-                  check</mat-icon>
-                <mat-icon matTooltip="delete" class="mat-ic"
-                  (click)="$event.stopPropagation(); deleteAsset(visit.assetId,visit.assetTitle);">
-                  delete_outline
-                </mat-icon>
-              </mat-cell>
-            </ng-container>
-            <mat-header-row *matHeaderRowDef="displayedColumns"></mat-header-row>
-            <mat-row class="example-element-row" (click)="navigateToAssetDetails(row.assetId)"
-              *matRowDef="let row; columns: displayedColumns;">
-            </mat-row>
-          </mat-table>
-          <div>
-            <mat-paginator [length]='totalCount' [pageSize]="5" [pageIndex]='pageNumber' (page)="pageChange($event)">
-            </mat-paginator>
-          </div>
-        </ng-template>
       </mat-card>
     </div>
   </div>
