+++ src/app/dashboard/dashboard.component.ts
@@ -14,18 +14,6 @@ export interface Year {
   data: string
 }
 
-interface Asset {
-  assetId: number;
-  isActive: string;
-  assetCodeImage: string;
-  assetCode: number;
-  assetImage: string;
-  assetTitle: string;
-  categoryName: string;
-  modelNumber: string;
-  companyAssetNo: string;
-}
-
 @Component({
   selector: 'app-dashboard',
   templateUrl: './dashboard.component.html',
@@ -70,26 +58,12 @@ export class DashboardComponent implements OnInit {
   CategoryMaintainanceTitle: any;
   InstallLocAssetTitle: any;
   installationlocation: any;
-  deleteAssetWithId: number;
-  isAlreadySubscribedToDialogUserActionService: boolean = false;
-
-  displayedColumns: string[] = ['assetTitle', 'addedBy', 'action'];
-
-  count: number;
-  pageNumber = 0;
-  totalCount = 0;
-  page: number = 0;
-  dataSource: MatTableDataSource<Asset> = new MatTableDataSource();
-  isNoRecordFound: boolean = false;
-
-  @ViewChild(MatPaginator) paginator: MatPaginator;
 
   constructor(
     private router: Router,
     private notificationService: NotificationService,
     private snackBar: MatSnackBar,
     private spinnerService: SpinnerService,
-    private assetmateService: AssetmateService,
     private dialogService: DialogService
   ) {
   }
@@ -110,42 +84,6 @@ export class DashboardComponent implements OnInit {
     this.categoryWiseMaintainance();
     this.installationLocWiseAsset();
 
-    this.getAllPendingVerificationAssets(this.pageNumber);
-
-  }
-
-  getAllPendingVerificationAssets(pageNo: number) {
-    this.spinnerService.setSpinnerVisibility(true);
-    this.assetmateService.getAllPendingVerificationAssets(pageNo).subscribe(res => {
-      this.spinnerService.setSpinnerVisibility(false);
-      if (res.status) {
-        if (res.currentPage == 0 && res.totalCount == 0) {
-          this.isNoRecordFound = true;
-        } else {
-          this.isNoRecordFound = false;
-        }
-        this.dataSource = res.Assets;
-        this.pageNumber = res.currentPage;
-        this.totalCount = res.totalCount;
-      } else {
-        this.showSnackBar(res.message);
-      }
-    },
-      error => {
-        this.spinnerService.setSpinnerVisibility(false);
-        this.showSnackBar("Something went wrong..!!");
-      }
-    );
-  }
-
-  pageChange(pageNo: any) {
-    this.page = pageNo.pageIndex;
-    this.getAllPendingVerificationAssets(this.page);
-  }
-
-  navigateToAssetDetails(assetId) {
-    let categoryId = 1;
-    this.router.navigate(['/assetmate/assetmate-details/' + categoryId + '/asset-details/' + assetId]);
   }
 
   years: Year[] = [
@@ -196,6 +134,7 @@ export class DashboardComponent implements OnInit {
       })
 
   }
+
   showSnackBar(message: string) {
     this.snackBar.open(message, '', { duration: 2000 });
   }
@@ -431,8 +370,6 @@ export class DashboardComponent implements OnInit {
         this.showSnackBar("Something went wrong..!!");
       })
   }
-
-
 
   /*********************************************Category Wise Assets Graph  ***************************************************/
 
@@ -722,64 +659,5 @@ export class DashboardComponent implements OnInit {
 
     window.open(doc.output('bloburl'), '_blank');
   }
-
-  verifiedAsset(assetId: number) {
-    let body = {
-      isVerified: 1
-    }
-    this.spinnerService.setSpinnerVisibility(true);
-    this.notificationService.verifyAsset(assetId, body).subscribe(res => {
-      this.spinnerService.setSpinnerVisibility(false);
-      if (res.status) {
-        if (res.currentPage == 0 && res.totalCount == 0) {
-          this.isNoRecordFound = true;
-        } else {
-          this.isNoRecordFound = false;
-        }
-        this.showSnackBar(res.message);
-        this.getAllPendingVerificationAssets(this.pageNumber);
-      } else {
-        this.spinnerService.setSpinnerVisibility(false);
-        this.showSnackBar(res.message);
-      }
-    },
-      error => {
-        this.spinnerService.setSpinnerVisibility(false);
-        this.showSnackBar("Something went wrong..!!");
-      })
-  }
-
-  deleteAsset(assetId: number, assetTitle: string) {
-    this.deleteAssetWithId = assetId;
-    let appDialogData: AppDialogData = {
-      visibilityStatus: true,
-      title: 'DELETE ASSET',
-      message: `Are your sure you want to delete asset "${assetTitle}" ?`,
-      positiveBtnLable: "Yes",
-      negativeBtnLable: "Cancel"
-    }
-    this.dialogService.setDialogVisibility(appDialogData);
-    if (!this.isAlreadySubscribedToDialogUserActionService) {
-      this.isAlreadySubscribedToDialogUserActionService = true;
-      this.dialogService.getUserDialogAction().subscribe(userAction => {
-        if (userAction == 0) {
-          //User has not performed any action on opened app dialog or closed the dialog;
-        } else if (userAction == 1) {
-          this.dialogService.setUserDialogAction(0);
-          //User has approved delete operation 
-          this.spinnerService.setSpinnerVisibility(true);
-          this.notificationService.deleteAsset(this.deleteAssetWithId).subscribe(res => {
-            this.spinnerService.setSpinnerVisibility(false);
-            this.showSnackBar(res.message);
-            this.getAllPendingVerificationAssets(this.pageNumber);
-          }, error => {
-            this.spinnerService.setSpinnerVisibility(false);
-            this.showSnackBar("Something went wrong..!!");
-          });
-        }
-      })
-    }
-  }
-
 
 }
