+++ src/app/modules/alert/alert-view/alert-view.component.ts
@@ -77,6 +77,7 @@ export class AlertViewComponent implements AfterViewInit, OnDestroy {
       if (res.alert) {
         if (res.currentPage == 0 && res.totalCount == 0) {
           this.isNoRecordFound = true;
+          this.showSnackBar(res.message);
         } else {
           this.isNoRecordFound = false;
         }
