+++ src/app/modules/manufacturer/manufacturer-view/add-manufacturer/add-manufacturer.component.ts
@@ -2,8 +2,6 @@ import { Component, OnInit, inject, Inject } from '@angular/core';
 import { Router } from '@angular/router';
 import { ManufacturerService } from '../../service/manufacturer.service';
 import { DataSharingService } from '../../../../public service/data-sharing.service';
-import { ToastrService } from 'ngx-toastr';
-import { NgxSpinnerService } from 'ngx-spinner';
 import { SpinnerService } from '../../../../public service/spinner.service';
 import { MatSnackBar, MatDialog } from '@angular/material';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
@@ -18,7 +16,6 @@ export class AddManufacturerComponent implements OnInit {
   formTitle: string = "Add Manufacturer";
   manufacturerData: any = {};
   cancelbtn = 0;
-
 
   constructor(private router: Router,
     private manufacturerService: ManufacturerService,
@@ -39,16 +36,6 @@ export class AddManufacturerComponent implements OnInit {
       this.isEdited = true;
       this.formTitle = `Edit Manufacturer`;
     }
-    // this.dataService.currentData.subscribe(res => {
-    //   if (res != null && res != "null" && res != "null") {
-    //     console.log("ngonit function", res);
-    //     this.manufacturerData.title = res.title;
-    //     this.manufacturerData.manufacturerId = res.manufacturerId;
-    //     this.isEdited = true;
-    //     this.formTitle = `Edit Manufacturer`;
-
-    //   }
-    // })
   }
 
   /*********************************************************** Add New Manufacturer *******************************************************************/
@@ -57,15 +44,16 @@ export class AddManufacturerComponent implements OnInit {
     this.spinnerService.setSpinnerVisibility(true);
     this.manufacturerService.addmanufacturer(value).subscribe(res => {
       this.spinnerService.setSpinnerVisibility(false);
-      this.showSnackBar(res.message);
-      this.dialog.closeAll();
-      //this.router.navigate(['/manufacturer']);
-
+      if (res.status) {
+        this.showSnackBar(res.message);
+        this.dialog.closeAll();
+      } else {
+        this.showSnackBar("Something went wrong..!!");
+      }
     },
       error => {
         this.spinnerService.setSpinnerVisibility(false);
         this.showSnackBar("Something went wrong..!!");
-
       })
   }
 
@@ -79,18 +67,21 @@ export class AddManufacturerComponent implements OnInit {
     this.spinnerService.setSpinnerVisibility(true);
     this.manufacturerService.editmanufacturer(this.manufacturerData.manufacturerId, value).subscribe(res => {
       this.spinnerService.setSpinnerVisibility(false);
-      this.showSnackBar(res.message);
-      this.dialog.closeAll();
-      //this.router.navigate(['/manufacturer']);
-
+      if (res.status) {
+        this.showSnackBar(res.message);
+        this.dialog.closeAll();
+      } else {
+        this.showSnackBar("Something went wrong..!!");
+      }
     },
       error => {
         this.spinnerService.setSpinnerVisibility(false);
         this.showSnackBar("Something went wrong..!!");
-
       })
   }
+
   /*********************************************************** Back to Manufacturer list *******************************************************************/
+
   backToList() {
     this.router.navigate(['/manufacturer']);
   }
