+++ src/app/modules/user/department-filter/department-filter.component.html
@@ -8,24 +8,29 @@
     </mat-form-field>
   </form>
   <div class="mar-29" *ngIf="isTreeDataReady">
-    <mat-tree [dataSource]="dataSource" [treeControl]="treeControl">
-      <!-- This is the tree node template for leaf nodes -->
-      <mat-tree-node class="mat-tree-node" *matTreeNodeDef="let node" matTreeNodePadding
-        (click)="handleCategoryTreeNodeClick(node)" class="mat-tree-node">
-        <!-- use a disabled button to provide padding for tree leaf -->
-        <button mat-icon-button disabled></button>
-        {{node.name}}
-      </mat-tree-node>
-      <!-- This is the tree node template for expandable nodes -->
-      <mat-tree-node *matTreeNodeDef="let node;when: hasChild" matTreeNodePadding class="mat-tree-node"
-        (click)="handleCategoryTreeNodeClick(node)" class="mat-tree-node">
-        <button mat-icon-button matTreeNodeToggle [attr.aria-label]="'toggle ' + node.name">
-          <mat-icon class="mat-icon-rtl-mirror">
-            {{treeControl.isExpanded(node) ? 'expand_more' : 'chevron_right'}}
-          </mat-icon>
-        </button>
-        {{node.name}}
-      </mat-tree-node>
-    </mat-tree>
+    <div *ngIf="isNoRecordFound; else recordFound" class="no-record-found-container">
+      <p>No Record Found</p>
+    </div>
+    <ng-template #recordFound>
+      <mat-tree [dataSource]="dataSource" [treeControl]="treeControl">
+        <!-- This is the tree node template for leaf nodes -->
+        <mat-tree-node class="mat-tree-node" *matTreeNodeDef="let node" matTreeNodePadding
+          (click)="handleCategoryTreeNodeClick(node)" class="mat-tree-node">
+          <!-- use a disabled button to provide padding for tree leaf -->
+          <button mat-icon-button disabled></button>
+          {{node.name}}
+        </mat-tree-node>
+        <!-- This is the tree node template for expandable nodes -->
+        <mat-tree-node *matTreeNodeDef="let node;when: hasChild" matTreeNodePadding class="mat-tree-node"
+          (click)="handleCategoryTreeNodeClick(node)" class="mat-tree-node">
+          <button mat-icon-button matTreeNodeToggle [attr.aria-label]="'toggle ' + node.name">
+            <mat-icon class="mat-icon-rtl-mirror">
+              {{treeControl.isExpanded(node) ? 'expand_more' : 'chevron_right'}}
+            </mat-icon>
+          </button>
+          {{node.name}}
+        </mat-tree-node>
+      </mat-tree>
+    </ng-template>
   </div>
 </mat-card>
