+++ src/app/modules/verify-asset/components/verify-asset-list/verify-asset-list.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { VerifyAssetListComponent } from './verify-asset-list.component';
+
+describe('VerifyAssetListComponent', () => {
+  let component: VerifyAssetListComponent;
+  let fixture: ComponentFixture<VerifyAssetListComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ VerifyAssetListComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(VerifyAssetListComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
