+++ src/app/modules/verify-asset/verify-asset-routing.module.ts
@@ -1,0 +1,10 @@
+import { Routes } from '@angular/router';
+import { VerifyAssetListComponent } from './components/verify-asset-list/verify-asset-list.component';
+
+export const VerifyAssetRoutes: Routes = [
+  {
+    path: "",
+    component: VerifyAssetListComponent
+  },
+
+];
